Add files via upload
This website has a online crime report managing system which is easily accessible by public and police department to create a general public awareness of the crime situation of their area . This application will also help general public to report a crime to law enforcement agency. Citizens can file a F.I.R from anywhere. It's a CSS and HTML based site with MySQL as a backend